+++ crates/hl7v2/src/transport/mllp/constants.rs
@@ -1,0 +1,8 @@
+/// MLLP start byte (vertical tab)
+pub const MLLP_START: u8 = 0x0B;
+
+/// MLLP end byte 1 (file separator)
+pub const MLLP_END_1: u8 = 0x1C;
+
+/// MLLP end byte 2 (carriage return)
+pub const MLLP_END_2: u8 = 0x0D;
